refactor:change structure of my app & start react-forms task
