+++ react/src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Header from './components/header/Header';
-import Home from './components/home/Home';
-import About from './components/about/About';
-import NotFound from './components/notFound/NotFound';
+import Header from './components/Header/Header';
+import Home from './pages/Home/Home';
+import About from './pages/About/About';
+import NotFound from './pages/NotFound/NotFound';
 import { Routes, Route } from 'react-router-dom';
 import './App.scss';
