+++ react/src/tests/Card.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import Card from '../components/home/Card/Card';
+import Card from '../components/Card/Card';
 
 const fields = ['title', 'year', 'country', 'genre', 'views'];
 beforeEach(() => {
