+++ react/src/tests/Search.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import Search from '../components/home/Search/Search';
+import Search from '../components/Search/Search';
 
 describe('Search component', () => {
   it('have search input', () => {
